% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{theia_check}
\alias{theia_check}
\alias{theia_check,TheiaTile-method}
\alias{theia_check,TheiaCollection-method}
\title{Check theia object}
\usage{
theia_check(object)

\S4method{theia_check}{TheiaTile}(object)

\S4method{theia_check}{TheiaCollection}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{TheiaTile}}

\item{object}{An object of class \linkS4class{TheiaCollection}}
}
\description{
Check an object of class Theia*
}
\section{Methods (by class)}{
\itemize{
\item \code{TheiaTile}: Check a single tile

\item \code{TheiaCollection}: Check a collection
}}

