% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{TheiaCart}
\alias{TheiaCart}
\title{Class to represent a theia cart}
\usage{
TheiaCart(cart.file, dest.dir)
}
\description{
Class to represent a theia cart
}
\section{Slots}{

\describe{
\item{\code{cart.file}}{'.meta4' cart file downloaded from Theia}

\item{\code{cart.infos}}{Informations about the cart}
}}

