% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{TheiaCart}
\alias{TheiaCart}
\title{Construct object of class TheiaCart}
\usage{
TheiaCart(cart.file, dest.dir)
}
\arguments{
\item{cart.file}{'.meta4' cart file downloaded from Theia}

\item{dest.dir}{Directory to store tiles}
}
\value{
An object of class \linkS4class{TheiaCart}
}
\description{
Construct an object of class TheiaCart.
}
