% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{TheiaCollection}
\alias{TheiaCollection}
\title{Class to represent a theia collection}
\usage{
TheiaCollection(from, dest.dir)
}
\description{
Class to represent a theia collection
}
\section{Slots}{

\describe{
\item{\code{from}}{Either a \code{character} giving the path to a cart file
('.meta4' file) or an object of class \linkS4class{TheiaQuery}}

\item{\code{cart.infos}}{Informations about the cart}

\item{\code{dest.dir}}{Directory to store tiles}
}}

