% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheiaCollection.R
\docType{class}
\name{TheiaCollection}
\alias{TheiaCollection}
\title{Class representing a collection of tiles from Theia}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TheiaCollection
}
\description{
Internal representation of a collection of tiles from Theia, storing their
locations, statuses and metadata.
}
\section{Fields}{

\describe{
\item{\code{dir.path}}{The path to the directory containing zip files}

\item{\code{tiles}}{A list of TheiaTile objects}

\item{\code{cart}}{An XML cart parsed from a 'meta4' file downloaded from Theia
website. Used only if Collection is created from a cart}

\item{\code{query}}{A TheiaQuery object, used only if collection is created from a
TheiaQuery object}
}}

\section{Methods}{

\describe{
   \item{\code{$new()}}{
     Create a new instance of the class
   }
   \item{\code{$download()}}{
     Download the tiles of the collection and check the resulting files
   }
   \item{\code{$check()}}{
     Check the tiles of the collection
   }
}
}

\keyword{datasets}
