% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TheiaQuery-class}
\alias{TheiaQuery-class}
\alias{TheiaQuery}
\title{Class to represent a theia query}
\description{
Class to represent a theia query
}
\section{Slots}{

\describe{
\item{\code{credentials}}{A list with login and password}

\item{\code{query}}{A list of search terms. See Details.}
}}

