% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{TheiaTile}
\alias{TheiaTile}
\title{Class to represent a theia single tile}
\usage{
TheiaTile(file.path, url, file.hash = NULL)
}
\description{
Class to represent a theia single tile
}
\section{Slots}{

\describe{
\item{\code{file.path}}{Path of the tile}

\item{\code{file.hash}}{md5 hash to check if the tile has been correctly downloaded}

\item{\code{url}}{URL form which the tile has been downloaded}

\item{\code{status}}{Status of the tile (whether it is present, checked)}
}}

