% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheiaTile.R
\name{TheiaTile}
\alias{TheiaTile}
\title{A tile from Theia}
\description{
Generate and manage a tile from Theia (download, check, load).
}
\section{Usage}{

\preformatted{
   t <- TheiaTile$new(file.path,
                      url,
                      file.hash)

   t$download(overwrite = FALSE)
   t$check()
   t$get_bands()
   t$read(bands)
   t$extract(overwrite = FALSE, dest.dir = NULL)
}
}

\section{Arguments}{


\describe{
   \item{t:}{A \code{TheiaTile} object}
   \item{file.path:}{The path to the zip file containing the tile}
   \item{url:}{The url to download the tile}
   \item{file.hash:}{The md5sum used to check the zip file}
   \item{auth:}{A character string giving the file path to Theia credentials.
   Or a \code{\link{TheiaAuth}} object}
   \item{overwrite:}{Overwrite existing tiles (default to `FALSE`)}
 }
}

\section{Details}{

   \code{TheiaTile$new(file.path, url, file.hash)} Create a new instance of
   the class

   \code{t$download(auth, overwrite = FALSE)} Download the tiles of the collection
   and check the resulting files

   \code{t$check()} Check the tiles of the collection

   \code{t$get_bands()} List bands available in the tile

   \code{t$read(bands)} Read band(s) from the zip file and returns a list of
   raster objects

   \code{t$extract(overwrite = FALSE, dest.dir = NULL)} Extract archive to
   dest.dir if supplied, or to the same directory as the archive otherwise
}

\examples{

file.path <- system.file("extdata", "dummy_tile.zip", package = "theiaR")
file.hash <- "7a6688ec0b711859a022cdbfd8b299f9"
url <-
 paste0("https://theia.cnes.fr/atdistrib/resto2/collections/SENTINEL2/",
        "5ea54171-f14b-5781-8d07-9be8e436410b/download")

mytile <- TheiaTile$new(file.path = file.path,
                        url       = url,
                        file.hash = file.hash,
                        tile.name = "SENTINEL2A_20190113-104826-809_L2A_T31TGK_D")

# Show status of tile
print(mytile$status)

\dontrun{
# Download the tiles
mytile$download()
}

}
