% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{theia_download}
\alias{theia_download}
\alias{theia_download,TheiaTile-method}
\title{Download theia object}
\usage{
theia_download(object, override = FALSE)

\S4method{theia_download}{TheiaTile}(object, override = FALSE)
}
\arguments{
\item{object}{}

\item{override}{Logical: download files that are already present? (default to
FALSE)}
}
\value{
Invisibly, a vector of downloaded file path
}
\description{
Download an object of class Theia*
}
\section{Methods (by class)}{
\itemize{
\item \code{TheiaTile}: Download a single tile
}}

