% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{TheiaCollection}
\alias{TheiaCollection}
\title{Create a \code{TheiaCollection} object}
\usage{
TheiaCollection(from, dest.dir)
}
\arguments{
\item{from}{Either a \code{character} giving the path to a cart file
('.meta4' file) or an object of class \linkS4class{TheiaQuery}}

\item{dest.dir}{Directory to store tiles}
}
\description{
Create a \code{TheiaCollection} object from a cart file or a
\linkS4class{TheiaQuery} object (not implemented yet)
}
