% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheiaCollection.R
\name{TheiaCollection}
\alias{TheiaCollection}
\title{A collection of tiles from Theia}
\description{
Generate and manage collection of tiles from Theia. This collection can be
created either from a cart file ('.meta4') downloaded from Theia website,
from a \code{\link{TheiaQuery}} object or from a list of
\code{\link{TheiaTile}} (not implemented yet).
}
\section{Usage}{

\preformatted{
   c <- TheiaCollection$new(cart.path = NULL,
                            tiles     = NULL,
                            query     = NULL,
                            dir.path  = NULL)

   c$download(auth, overwrite = FALSE)
   c$check()
   c$get_bands()
   c$status
}
}

\section{Arguments}{


\describe{
   \item{c:}{A \code{TheiaCollection} object}
   \item{dir.path:}{The path to the directory containing zip files}
   \item{tiles:}{A list of TheiaTile objects}
   \item{cart:}{An XML cart parsed from a 'meta4' file downloaded from Theia}
   website. Used only if Collection is created from a cart
   \item{query:}{A TheiaQuery object, used only if collection is created
   from a TheiaQuery object. Can also be a list with search terms. In this
   case, it will create a `TheiaQuery` object from it. See
   \code{\link{TheiaQuery}} for details on query syntax}
   \item{auth:}{A cahracter string giving the file pathto Theia credentials.
   Or a \code{\link{TheiaAuth}} object}
   \item{overwrite:}{Overwrite existing tiles (default to `FALSE`}
 }
}

\section{Details}{

   \code{TheiaCollection$new()} Create a new instance of the class

   \code{c$download(overwrite = FALSE)} Download the tiles of the collection
   and check the resulting files

   \code{$ccheck()} Check the tiles of the collection

   \code{c$get_bands()} List bands available in each tile

   \code{c$status} Return the status of each tile of teh collection

   \code{c$read(bands)} Read band(s) from the zip files and returns a list of
   raster objects
}

