% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheiaTile.R
\docType{class}
\name{TheiaTile}
\alias{TheiaTile}
\title{Class representing a single tile from Theia}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TheiaTile
}
\description{
Internal representation of a tile from Theia, storing its location, status
and metadata.
}
\section{Fields}{

\describe{
\item{\code{file.path}}{The path to the zip file containing the tile}

\item{\code{file.hash}}{The hash of the file, used to check if the file is correctly
downloaded}

\item{\code{url}}{The url to download the tile from Theia website}

\item{\code{status}}{A list giving the status of the tile: whether it is downloaded,
checked, correctly downloaded}
}}

\section{Methods}{

\describe{
   \item{\code{$new()}}{
     Create a new instance of the class
   }
   \item{\code{$download()}}{
     Download the tile and check the resulting file
   }
   \item{\code{$check()}}{
     Check the tile
   }
}
}

\keyword{datasets}
