% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{TheiaTile}
\alias{TheiaTile}
\title{Construct object of class TheiaTile}
\usage{
TheiaTile(file.path, url, file.hash = NULL)
}
\arguments{
\item{file.path}{Path of file}

\item{url}{URL}

\item{file.hash}{file hash}
}
\value{
An object of class \linkS4class{TheiaTile}
}
\description{
Construct an object of class TheiaTile.
}
