% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheiaTile.R
\name{TheiaTile}
\alias{TheiaTile}
\title{A tile from Theia}
\description{
Generate and manage a tile from Theia (download, check, load).
}
\section{Usage}{

\preformatted{
   t <- TheiaTile$new(file.path,
                      url,
                      file.hash)

   t$download(override = FALSE)
   t$check()
   t$get_bands()
   t$read(bands)
}
}

\section{Arguments}{


\describe{
   \item{t:}{A \code{TheiaTile} object}
   \item{file.path:}{The path to the zip file containing the tile}
   \item{url:}{The url to download the tile}
   \item{file.hash:}{The md5sum used to check the zip file}
   \item{override:}{Override existing tiles (default to `FALSE`)}
 }
}

\section{Details}{

   \code{TheiaTile$new(file.path, url, file.hash)} Create a new instance of
   the class

   \code{t$download(override = FALSE)} Download the tiles of the collection
   and check the resulting files

   \code{t$check()} Check the tiles of the collection
   \code{t$get_bands()} List bands available in the tile
   \code{t$read(bands)} Read band(s) from the zip file
}

